\docType{package}
\name{ATOdig-package}
\alias{ATO.dig}
\alias{ATOdig-package}
\title{Access the ATO Excel based data from R}
\description{
  A package to access the ATO Excel based data from R
}
\details{
  \tabular{ll}{ Package: \tab ATOdig\cr Type: \tab
  Package\cr Version: \tab 0.1-3\cr Date: \tab
  2013-06-03\cr License: \tab LGPL 2.1\cr }
}
\author{
  Jean-Michel Perraud
  \email{jean-michel.perraud_at_csiro.au}
}
\keyword{.NET}
\keyword{ATO}
\keyword{Mono}
\keyword{package}
\keyword{rClr}

