\name{extractSeries}
\alias{extractSeries}
\title{find the categories in the table extracted from the ATO excel file}
\usage{
  extractSeries(d)
}
\arguments{
  \item{d}{the character matrix}
}
\value{
  a list of data frames, with the series for ATO data.
}
\description{
  find the categories in the table extracted from the ATO
  excel file
}

