\name{findCategories}
\alias{findCategories}
\title{find the categories in the table extracted from the ATO excel file}
\usage{
  findCategories(str_mat)
}
\arguments{
  \item{str_mat}{the character matrix}
}
\value{
  nothing is returned by this function
}
\description{
  find the categories in the table extracted from the ATO
  excel file
}

